moves tests around based on pr feedback
